Create eth_coverter.sol
EthValConverter contract has three state variables: weiVal, gweiVal, and etherVal.

The EthConverterfunction is  payable, which means that it can receive ether from the user. When the function is called, it sets the weiVal state variable to the amount of ether sent by the user in wei.

It then converts this value to gwei (which is one billionth of an ether) by dividing it by 10^9, and stores the result in the gweiVal state variable.

Finally, it converts the value to ether by dividing it by the 1 ether constant, which represents one ether in wei. The result is stored in the etherVal state variable.