Create MyToken.sol
This is a simple Solidity smart contract that represents a token called EDUCA with the token abbreviation EDU. The contract has a few basic features that allow users to mint new tokens and burn existing tokens. Here's an overview of the contract:

The contract has three public state variables: tokenname, tokenabbrv, and totalsupply. These variables store the name, abbreviation, and total supply of the token, respectively.

The contract has a mapping called balances that maps addresses to their respective token balances. When the contract is deployed, all token balances are set to 0.

The mint function takes two input parameters: an address and a value. When called, the function increases the total supply by the value and increases the balance of the given address by the same amount.

The burn function works the opposite of the mint function. It also takes an address and a value as input parameters. When called, the function decreases the total supply by the value and decreases the balance of the given address by the same amount, but only if the balance of the given address is greater than or equal to the value.